[6X]gprecoverseg rebalance is failed with timeout while promoting mirrors
Currently gprecoverseg -r waits for a maximum of 30 seconds for preferred primaries to be promoted. This may not be enough and results in an exception

gprecoverseg rebalance workflow

All the segments are up and running but the roles are switched
User runs gprecoverseg -r
gprecoverseg shuts down all the acting primaries(preferred mirrors)
gprecoverseg triggers fts probe
gprecoverseg sends a test query (CREATE TEMP TABLE) to start a distributed transaction to test if all the preferred primaries were promoted
gprecoverseg waits for a max of 30 secs and checks if the CREATE TABLE query succeeded, else raises timeout exception
gprecoverseg starts the stopped preferred mirrors
The problem is that in some cases, the preferred primary may not be promoted in 30 seconds which results in an exception. For now we have decided to increase the timeout to 10 mins which should be enough for most use cases.
Eventually we will revisit this logic and make it a bit more robust.