Modify cardinality estimation for local HashAggregate
To ensure accurate cardinality estimation for local aggregation, we
consider the following scenarios:
 - The tuples of local agg fit within memory.
 - We handle worst-case scenarios where the tuples of the grouping key
   are distributed across segments.
To maintain accurate cardinality, the local agg's cardinality should not
exceed the NDV of the grouping key in global aggs (upper bound for the
number of output rows). To achieve this, we multiply the global agg's
cardinality of the grouping key with the number of segments.  This
approach helps to maintain the cardinality for local aggregation across
both best and worst-case scenarios.
Current approach:
	rows = (rows + num_rows_outer) / 2.0;

New approach:
	rows = rows * pcmgpdb->UlHosts();

Addressed unit/regression testcase failures