Adjust per-segment row calculations when the child is replicated
Orca incorrectly sets the operator's per-segment rows when its child is
replicated.
for example :
|--CPhysicalMotionHashDistribute HASHED: [ CScalarIdent “a” (8) rows:10000000]
+--CPhysicalTableScan “rep” (“rep”)   rows:10000000
Note: number_of_segments=3

In the above example, for the operator CPhysicalMotionHashDistribute, we
incorrectly set the per-segment number of rows to 3333333 instead of
10000000. This incorrect setting is the reason for the lower cost of the
hash join alternative (redistribute, redistribute).
This PR also addresses the issue of adjusting per-segment row
calculations when the operator's child is replicated.