Add new optmization request to HashInnerJoin
Issue:
Orca isn't generating the join alternative where the replicated table is
on the outer and the randomly distributed table is on the inner with no
motions.

RCA:
During hash join, we calculate the required distribution for the n-th
child at CPhysicalHashJoin::Ped(). It will create three optimization
requests for joining the children.

1. Req(redistribute, redistribute) <outer child, innerchild>
2. Req(hashed/non-singleton, broadcast) <outer child, innerchild>
3. Req(singleton, singleton) <outer child, innerchild>

However, we haven't created the necessary optimization request for
scenarios where the outer table is replicated, and the inner table is
randomly distributed. This omission leads to unnecessary enforcers being
introduced. Orca typically selects one of the above three alternatives
based on cost, even when an enforcer is not required when the inner
table is randomly distributed.
In this PR, I'm introducing a fourth optimization request to
HashInnerJoin to address this specific scenario.
4. Req(Replicated, Non-Replicated)