Introduce a new Optimization Request(broadcast, Non-Singleton) for InnerHashJoin
Orca fails to generate the most efficient join alternative when the
replicated table is on the outer side, and the randomly distributed
table is on the inner side. Consequently, we are consistently opting for
a plan that involves unnecessary motions on join children.

RCA:
During hash join, we calculate the required distribution for the n-th
child at CPhysicalHashJoin::Ped(). It will create three optimization
requests for joining the children.

Req(redistribute, redistribute) <outer child, innerchild>
Req(hashed/non-singleton, broadcast) <outer child, innerchild>
Req(singleton, singleton) <outer child, innerchild>

However, we haven't created the necessary optimization request for
scenarios where the outer table is replicated, and the inner table is
randomly distributed. This omission leads to unnecessary enforcers being
introduced on join children.

So the objective is to allow the plan shape of an outer replicated table
to hash join with an inner distributed table without requiring
motion. This is achieved by only allowing the (broadcast, non-singleton)
request to cases where the outer table is already replicated.

The PR changes consist of the following:
  - Inclusion of a fourth optimization request (Broadcast,
    Non-Singleton) in HashInnerJoin to handle cases where the outer table
    is replicated, and the inner table has random distribution.
  - Restricting the new optimization request (Broadcast, Non-Singleton)
    to cases where the join's outer child is replicated. Specifically,
    we prohibit the enforce under the following conditions:
    1. If it is a broadcast outer request, indicated by the
       fAllowedEnforced = false flag (set to true for the broadcast
       inner request).
    2.If the input distribution spec is not already replicated, or if
      the input distribution is replicated due to an enforcer introduced
      by another request.  An example of the latter is when a broadcast
      motion is applied to a non-replicated table on the inner side due
      to a commutativity transformation.

   By prohibiting the enforcer, we disregard the optimization context,
   irrespective of whether the input distribution spec meets the
   requested spec. This leaves the optimization context without a best
   group expression, causing the optimization context to be bypassed by
   the cost model.

   Note: We are aware the changes proposed by this PR circumvents the
   cost model.  And in certain scenarios, broadcasting a non-replicated
   outer may yield better performance, but isn't permitted by this PR.
   This approach is deliberately chosen to minimize disruptions to
   current plans and introduce a plan alternative only under particular
   circumstances.

   Eg. two small distributed tables with equal tuples joining on the
   non-distribution key

   In this case, broadcasting the outer table is better than
   broadcasting the inner one.  Since we always hash the inner table,
   the hashing cost of a distributed table is lower than that of a
   replicated table. However, as detailed by the reasoning above,
   broadcasting a distributed outer isn't permitted by this PR.

Testcases:
- Added regression cases to cover scenarios
   - replicated table as outer child and random/hash table as inner child
   - random/hash table as outer child and replicated table as inner child
   - join of two replicated tables
   - join of two random/hash tables
   - join of universal data with replicated table
Note: I noticed that there are already existing MDPs covering the
scenarios, so I didn't add any new MDP test cases.