[ORCA] Fix bug in per-segment rows calculation when a child is replicated
Orca's behavior in setting the operator's per-segment rows is incorrect
when the child is replicated. As a result, the hash join often chooses
the [hash, hash]alternative, even when better alternatives are
available.

Example:
create table rep as select generate_series(1,10000) a distributed
replicated;
create table rand as select generate_series(1,100) a
distributed randomly;
analyze rep, rand;
explain select * from rand t1 join rep t2 on t1.a = t2.a;
PhysicalPlan:
+--CPhysicalMotionGather(coordinator)
   +--CPhysicalInnerHashJoin
      |--CPhysicalMotionHashDistribute HASHED
      |  +--CPhysicalTableScan "rep" ("rep")
      |--CPhysicalMotionHashDistribute HASHED
      |  +--CPhysicalTableScan "rand" ("rand")
      +--CScalarCmp (=)
         |--CScalarIdent "a" (0)
         +--CScalarIdent "a" (8)

In the above example, with regards to the operator
CPhysicalMotionHashDistribute over the 'rep' table, we are incorrectly
setting  the per-segment number of rows to 3333 instead of 10000. This
incorrect setting is the reason for the lower cost of the hash join
alternative (redistribute, redistribute). This PR addresses the issue of
adjusting per-segment row calculations when the operator's child is
replicated. After applying the fix, it was noticed that the query's
execution time decreased compared to the previous performance